requirements.txt
cffi==1.15.1
cryptography==38.0.4
dnspython==2.2.1
filelock==3.8.2
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
py3-validate-email==1.0.5.post1
pycparser==2.21
python-dotenv==0.21.0
